main: - Supplier tg_id from int to bigint

+++ src/domain/suppliers/adapters/orm.py
@@ -7,7 +7,7 @@ from src.domain.suppliers.domain import Supplier
 supplier = sa.Table(
     'supplier',
     mapper.metadata,
-    sa.Column('tg_id', sa.Integer, primary_key=True),
+    sa.Column('tg_id', sa.BigInteger, primary_key=True),
     sa.Column('title', sa.String(255), nullable=False),
     sa.Column('alias', sa.String(255), nullable=False),
 )
